+++ src/CrossValidation/HoldOut.php
@@ -3,7 +3,9 @@
 namespace Rubix\ML\CrossValidation;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Estimator;
 use Rubix\ML\Datasets\Labeled;
+use Rubix\ML\Other\Helpers\Params;
 use Rubix\ML\CrossValidation\Metrics\Metric;
 use InvalidArgumentException;
 
@@ -57,13 +59,20 @@ class HoldOut implements Validator
      * @param  \Rubix\ML\Learner  $estimator
      * @param  \Rubix\ML\Datasets\Labeled  $dataset
      * @param  \Rubix\ML\CrossValidation\Metrics\Metric  $metric
+     * @throws \InvalidArgumentException
      * @return float
      */
     public function test(Learner $estimator, Labeled $dataset, Metric $metric) : float
     {
+        if (!in_array($estimator->type(), $metric->compatibility())) {
+            throw new InvalidArgumentException(Params::shortName($metric)
+                . ' is not compatible with '
+                . Estimator::TYPES[$estimator->type()] . 's.');
+        }
+
         list($testing, $training) = $this->stratify
             ? $dataset->stratifiedSplit($this->ratio)
-            : $dataset->randomize()->split($this->ratio);
+            : $dataset->split($this->ratio);
 
         $estimator->train($training);