+++ src/CrossValidation/KFold.php
@@ -3,8 +3,10 @@
 namespace Rubix\ML\CrossValidation;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Estimator;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Other\Helpers\Stats;
+use Rubix\ML\Other\Helpers\Params;
 use Rubix\ML\CrossValidation\Metrics\Metric;
 use InvalidArgumentException;
 
@@ -64,13 +66,20 @@ class KFold implements Validator
      * @param  \Rubix\ML\Learner  $estimator
      * @param  \Rubix\ML\Datasets\Labeled  $dataset
      * @param  \Rubix\ML\CrossValidation\Metrics\Metric  $metric
+     * @throws \InvalidArgumentException
      * @return float
      */
     public function test(Learner $estimator, Labeled $dataset, Metric $metric) : float
     {
+        if (!in_array($estimator->type(), $metric->compatibility())) {
+            throw new InvalidArgumentException(Params::shortName($metric)
+                . ' is not compatible with '
+                . Estimator::TYPES[$estimator->type()] . 's.');
+        }
+
         $folds = $this->stratify
             ? $dataset->stratifiedFold($this->k)
-            : $dataset->randomize()->fold($this->k);
+            : $dataset->fold($this->k);
 
         $score = 0.;