+++ src/CrossValidation/Metrics/Accuracy.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Metrics;
 
+use Rubix\ML\Estimator;
 use InvalidArgumentException;
 
 /**
@@ -26,6 +27,18 @@ class Accuracy implements Metric
         return [0., 1.];
     }
 
+    /**
+     * The estimator types that this metric is compatible with.
+     * 
+     * @return int[]
+     */
+    public function compatibility() : array
+    {
+        return [
+            Estimator::CLASSIFIER,
+            Estimator::DETECTOR,
+        ];
+    }
 
     /**
      * Score a set of predictions.