+++ src/CrossValidation/Metrics/F1Score.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Metrics;
 
+use Rubix\ML\Estimator;
 use InvalidArgumentException;
 
 /**
@@ -23,6 +24,19 @@ class F1Score implements Metric
     public function range() : array
     {
         return [0., 1.];
+    }
+
+    /**
+     * The estimator types that this metric is compatible with.
+     * 
+     * @return int[]
+     */
+    public function compatibility() : array
+    {
+        return [
+            Estimator::CLASSIFIER,
+            Estimator::DETECTOR,
+        ];
     }
 
     /**