+++ src/CrossValidation/Metrics/Informedness.php
@@ -3,8 +3,6 @@
 namespace Rubix\ML\CrossValidation\Metrics;
 
 use Rubix\ML\Estimator;
-use Rubix\ML\Datasets\Dataset;
-use Rubix\ML\Datasets\Labeled;
 use InvalidArgumentException;
 
 /**
@@ -34,6 +32,19 @@ class Informedness implements Metric
     public function range() : array
     {
         return [0., 1.];
+    }
+
+    /**
+     * The estimator types that this metric is compatible with.
+     * 
+     * @return int[]
+     */
+    public function compatibility() : array
+    {
+        return [
+            Estimator::CLASSIFIER,
+            Estimator::DETECTOR,
+        ];
     }
 
     /**